Correcion

+++ Guia_1/p3.py
@@ -1,5 +1,6 @@
+# Ejercicio 3
 print("Conversor de Grados Celcius a Fahrenheit")
-x = int(input("Indique la temperatura en °C: "))
+x = float(input("Indique la temperatura en °C: "))
 y = (x * 9/5) + 32
 
 print("La temperatura en °F es de: ", y , "°F")
